% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gev_prob}
\alias{cpp_gev_prob}
\title{Informative prior for GEV parameters constructed on the probability scale.}
\usage{
cpp_gev_prob(x, ppars)
}
\description{
Informative prior for GEV parameters constructed on the probability scale.
}
