% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gev_quant}
\alias{cpp_gev_quant}
\title{Informative prior for GEV parameters constructed on the quantile scale.}
\usage{
cpp_gev_quant(x, ppars)
}
\description{
Informative prior for GEV parameters constructed on the quantile scale.
}
