% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_prior_xptr}
\alias{create_prior_xptr}
\title{Create external pointer to a C++ function for a log-prior}
\usage{
create_prior_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ function for a log-prior
}
