% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps.R
\name{dgaps_post}
\alias{dgaps_post}
\title{Random sampling from D-gaps posterior distribution}
\usage{
dgaps_post(
  data,
  thresh,
  D = 1,
  n = 1000,
  inc_cens = TRUE,
  alpha = 1,
  beta = 1,
  param = c("logit", "theta"),
  use_rcpp = TRUE
)
}
\arguments{
\item{data}{A numeric vector or numeric matrix of raw data.  If \code{data}
  is a matrix then the log-likelihood is constructed as the sum of
  (independent) contributions from different columns. A common situation is
  where each column relates to a different year.

  If \code{data} contains missing values then
  \code{\link[exdex]{split_by_NAs}} is used to divide the data further into
  sequences of non-missing values, stored in different columns in a matrix.
  Again, the log-likelihood is constructed as a sum of contributions from
  different columns.}

\item{thresh}{A numeric scalar.  Extreme value threshold applied to data.}

\item{D}{A numeric scalar.  The censoring parameter \eqn{D}, as defined in
Holesovsky and Fusek (2020). Threshold inter-exceedances times that are
not larger than \code{D} units are left-censored, occurring with
probability \eqn{\log(1 - \theta e^{-\theta d})}{log(1 - exp(-\theta d))},
where \eqn{d = q D} and \eqn{q} is the probability with which the
threshold \eqn{u} is exceeded.}

\item{n}{A numeric scalar. The size of posterior sample required.}

\item{inc_cens}{A logical scalar indicating whether or not to include
contributions from right-censored inter-exceedance times, relating to the
first and last observations. It is known that these times are greater
than or equal to the time observed.
If \code{data} has multiple columns then there will be right-censored
first and last inter-exceedance times for each column. See also the
\strong{Details} section of \code{\link[exdex]{dgaps}}.}

\item{alpha, beta}{Positive numeric scalars.  Parameters of a
beta(\eqn{\alpha}, \eqn{\beta}) prior for \eqn{\theta}.}

\item{param}{A character scalar.  If \code{param = "logit"} (the default)
then we simulate from the posterior distribution of
\eqn{\phi = \log(\theta / (1-\theta))}{\phi = log(\theta / (1-\theta))}
and then transform back to the
\eqn{\theta}-scale.  If \code{param = "theta"} then we simulate
directly from the posterior distribution of \eqn{\theta}, unless
the sample D-gaps are all equal to zero or all positive, when we revert
to \code{param = "logit"}.  This is to avoid the possibility of sampling
directly from a posterior with mode equal to 0 or 1.}

\item{use_rcpp}{A logical scalar.  If \code{TRUE} (the default) the
rust function \code{\link[rust]{ru_rcpp}} is used for
posterior simulation.  If \code{FALSE} the (slower) function
\code{\link[rust]{ru}} is used.}
}
\value{
An object (list) of class \code{"evpost"}, which has the same
  structure as an object of class \code{"ru"} returned from
  \code{\link[rust]{ru}}.
  In addition this list contains
  \itemize{
    \item{\code{call}:} The call to \code{dgaps()}.
    \item{\code{model}:} The character scalar \code{"dgaps"}.
    \item{\code{thresh}:} The argument \code{thresh}.
    \item{\code{ss}:} The sufficient statistics for the D-gaps likelihood,
      as calculated by \code{\link[exdex]{dgaps_stat}}.
  }
}
\description{
Uses the \code{\link[rust]{rust}} package to simulate from the posterior
distribution of the extremal index \eqn{\theta} based on the D-gaps model
for threshold interexceedance times of Holesovsky and Fusek (2020). We refer
to this as the \eqn{D}-gaps model, because it uses a tuning parameter
\eqn{D}, whereas the related \eqn{K}-gaps model of Suveges and Davison
(2010) has a tuning parameter \eqn{K}.
}
\details{
A beta(\eqn{\alpha}, \eqn{\beta}) prior distribution is used for
  \eqn{\theta} so that the posterior from which values are simulated is
  proportional to
  \deqn{\theta ^ {2 N_1 + \alpha - 1}
    (1 - \theta e^{-\theta d}) ^ {N_0 + \beta - 1}
    \exp\{- \theta q (I_0 T_0 + \cdots + I_N T_N)\}.}{%
    \theta ^ (2 N_1 + \alpha - 1) *
    (1 - \theta e^{-\theta d}) ^ (N_0 + \beta - 1) *
    exp(- \theta q (I_0 T_0 + \dots + I_N T_N)).}
  See \code{\link[exdex]{dgaps_stat}} for a description of the variables
  involved in the contribution of the likelihood to this expression.

  The \code{\link[rust]{ru}} function in the \code{\link[rust]{rust}}
  package simulates from this posterior distribution using the
  generalised ratio-of-uniforms distribution.  To improve the probability
  of acceptance, and to ensure that the simulation will work even in
  extreme cases where the posterior density of \eqn{\theta} is unbounded as
  \eqn{\theta} approaches 0 or 1, we simulate from the posterior
  distribution of
  \eqn{\phi = \log(\theta / (1-\theta))}{\phi = log(\theta / (1-\theta))}
  and then transform back to the \eqn{\theta}-scale.
}
\examples{
# Newlyn sea surges

thresh <- quantile(newlyn, probs = 0.90)
d_postsim <- dgaps_post(newlyn, thresh)
plot(d_postsim)

### Cheeseboro wind gusts

d_postsim <- dgaps_post(exdex::cheeseboro, thresh = 45, D = 3)
plot(d_postsim)
}
\references{
Holesovsky, J. and Fusek, M. Estimation of the extremal index
  using censored distributions. Extremes 23, 197-213 (2020).
  \doi{10.1007/s10687-020-00374-3}

Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221. \doi{10.1214/09-AOAS292}
}
\seealso{
\code{\link[rust]{ru}} for the form of the object returned by
  \code{dgaps_post}.

\code{\link{kgaps_post}} for Bayesian inference about the
  extremal index \eqn{\theta} using the \eqn{K}-gaps model.
}
