% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gev_logpost_xptr}
\alias{gev_logpost_xptr}
\title{Create external pointer to a C++ GEV log-posterior function}
\usage{
gev_logpost_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ GEV log-posterior function
}
