% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_priors.R
\name{gev_prob}
\alias{gev_prob}
\title{Informative prior for GEV parameters (\eqn{\mu, \sigma, \xi}) constructed on the
probability scale.}
\usage{
gev_prob(pars, quant, alpha, min_xi = -Inf, max_xi = Inf, trendsd = 0)
}
\arguments{
\item{pars}{A numeric vector of length 3.
GEV parameters (\eqn{\mu, \sigma, \xi}).}

\item{quant}{A numeric vector of length 3.}

\item{alpha}{A numeric vector of length 4.}

\item{min_xi}{A numeric scalar.  Prior lower bound on \eqn{\xi}.}

\item{max_xi}{A numeric scalar.  Prior upper bound on \eqn{\xi}.}

\item{trendsd}{Has no function other than to achieve compatability with
function in the evdbayes package.}
}
\value{
The log of the prior density.
}
\description{
For information about this and other priors see \code{\link{set_prior}}.
}
