% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes.R
\docType{data}
\name{gom}
\alias{gom}
\title{Storm peak significant wave heights from the Gulf of Mexico}
\format{
A vector containing 315 observations.
}
\source{
Oceanweather Inc. (2005) GOMOS -- Gulf of Mexico hindcast study.
}
\usage{
gom
}
\description{
A numeric vector containing 315 hindcasts of storm peak significant wave
heights, metres, from 1900 to 2005 at an unnamed location in the Gulf
of Mexico.
}
\references{
Northrop, P. J., Attalides, N. and Jonathan, P. (2017)
  Cross-validatory extreme value threshold selection and uncertainty
  with application to ocean storm severity.
  \emph{Journal of the Royal Statistical Society Series C: Applied
  Statistics}, \strong{66}(1), 93-120.
  \url{https://doi.org/10.1111/rssc.12159}
}
\keyword{datasets}
