% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gp}
\alias{dgp}
\alias{gp}
\alias{pgp}
\alias{qgp}
\alias{rgp}
\title{The Generalised Pareto Distribution}
\usage{
dgp(x, scale = 1, shape = 0, log = FALSE)

pgp(q, scale = 1, shape = 0, lower_tail = TRUE)

qgp(p, scale = 1, shape = 0, lower_tail = TRUE)

rgp(n, scale = 1, shape = 0)
}
\arguments{
\item{x, q}{Numeric vectors of quantiles.  All elements of \code{x}
and \code{q} must be non-negative.}

\item{scale, shape}{Numeric vectors, of the same length.
Scale and shape parameters.
All elements of \code{scale} must be positive.
In \code{dgp}, \code{pgp} and \code{qgp} these vectors can have
length > 1, but only if \code{length(x)[dgp]}, \code{length(q)[pgp]},
or \code{length(p)[qgp]} has length one.
Otherwise, they must have length one.}

\item{log}{A logical scalar.  If TRUE the log-density is returned.}

\item{lower_tail}{A logical scalar.  If TRUE (default), probabilities
are P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities in (0,1).}

\item{n}{Numeric scalar.  The number of observations to be simulated.}
}
\value{
\code{dgp} gives the density function, \code{pgp} gives the
  distribution function, \code{qgp} gives the quantile function,
  and \code{rgp} generates random deviates.
}
\description{
Density function, distribution function, quantile function and
random generation for the generalised Pareto (GP) distribution.
}
\details{
The distribution function of a GP distribution with parameters
 \code{scale} = \eqn{\sigma} (>0) and \code{shape} = \eqn{\xi} is
 \deqn{F(x) = 1 - [1 + \xi x / \sigma] ^ (-1/\xi) }
 for \eqn{1 + \xi x / \sigma > 0}.  If \eqn{\xi = 0} the
 distribution function is defined as the limit as \eqn{\xi} tends to zero.
 The support of the distribution depends on \eqn{\xi}: it is
 \eqn{x >= 0} for \eqn{\xi >= 0};
 and \eqn{0 <= x <= - \sigma / \xi} for \eqn{\xi < 0}.  Note that
 if \eqn{\xi < -1} the GP density function becomes infinite as \eqn{x}
 approaches \eqn{-\sigma/\xi}.

 See
 \url{https://en.wikipedia.org/wiki/Generalized_Pareto_distribution}
  for information.
}
\examples{
dgp(0:4, 0.5, 0.8)
dgp(1:6, 0.5, -0.2, log = TRUE)
dgp(1, c(1, 2), c(-0.2, 0.4))

pgp(0:4, 0.5, 0.8)
pgp(1:6, 0.5, -0.2)
pgp(1, c(1, 2), c(-0.2, 0.4))
pgp(7, c(1, 2), c(-0.2, 0.4))
pgp(-3, c(1, 2), c(-0.2, 0.4))

qgp((1:9)/10, 0.5, 0.8)
qgp(0.5, c(0.5, 1), c(-0.5, 0.5))

p <- (1:9)/10
pgp(qgp(p, 1, 2, 0.8), 2, 0.8)

rgp(6, 0.5, 0.8)
}
\references{
Pickands, J. (1975) Statistical inference using extreme
  order statistics. \emph{Annals of Statistics}, \strong{3}, 119-131.
  \url{http://dx.doi.org/10.1214/aos/1176343003}

Coles, S. G. (2001) \emph{An Introduction to Statistical
  Modeling of Extreme Values}, Springer-Verlag, London.
  \url{http://dx.doi.org/10.1007/978-1-4471-3675-0}
}

