% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_xptr}
\alias{loglik_xptr}
\title{Create external pointer to a C++ function for logf}
\usage{
loglik_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ function for logf
}
