% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpost_xptr}
\alias{logpost_xptr}
\title{Create external pointer to a C++ log-posterior function}
\usage{
logpost_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ log-posterior function
}
