% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logprior_xptr}
\alias{logprior_xptr}
\title{Create external pointer to a C++ prior function}
\usage{
logprior_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ prior function
}
