% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{os_logpost_xptr}
\alias{os_logpost_xptr}
\title{Create external pointer to a C++ OS log-posterior function}
\usage{
os_logpost_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ OS log-posterior function
}
