% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes.R
\docType{data}
\name{oxford}
\alias{oxford}
\title{Annual Maximum Temperatures at Oxford}
\format{
A vector containing 80 observations.
}
\source{
Tabony, R. C. (1983) Extreme value analysis in meteorology.
 \emph{The Meteorological Magazine}, \strong{112}, 77-98.
}
\usage{
oxford
}
\description{
A numeric vector containing annual maximum temperatures, in degrees
Fahrenheit, from 1901 to 1980 at Oxford, England.
}
\keyword{datasets}
