% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes.R
\docType{data}
\name{portpirie}
\alias{portpirie}
\title{Annual Maximum Sea Levels at Port Pirie, South Australia}
\format{
A numeric vector containing 65 observations.
}
\source{
Coles, S. G. (2001) \emph{An Introduction to Statistical Modelling
  of Extreme Values}. London: Springer.
  \url{https://doi.org/10.1007/978-1-4471-3675-0}
}
\usage{
portpirie
}
\description{
A numeric vector of length 65 containing annual maximum sea levels,
in metres, from 1923 to 1987 at Port Pirie, South Australia.
}
\keyword{datasets}
