% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{print.evpost}
\alias{print.evpost}
\title{Print method for objects of class "evpost"}
\usage{
\method{print}{evpost}(x, ...)
}
\arguments{
\item{x}{An object of class "evpost", a result of a call to
\code{\link{rpost}}, \code{\link{rpost_rcpp}}, \code{\link{kgaps_post}}
or \code{\link{dgaps_post}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
The argument \code{x} is returned, invisibly.
}
\description{
Print method for objects of class "evpost"
}
\details{
\code{print.evpost} just prints the original function call, to
  avoid printing a huge list.
}
\examples{
# Newlyn sea surges

thresh <- quantile(newlyn, probs = 0.90)
k_postsim <- kgaps_post(newlyn, thresh)
k_postsim
}
\seealso{
\code{\link{plot.evpost}} for a diagnostic plot.
}
