% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Daily Aggregate Rainfall}
\format{
A vector containing 20820 observations.
}
\source{
Unknown
}
\usage{
rainfall
}
\description{
A numeric vector of length 20820 containing daily aggregate rainfall
observations, in millimetres, recorded at a rain gauge in England
over a 57 year period, beginning on a leap year. Three of these years
contain only missing values.
}
\keyword{datasets}
