% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes-internal.R
\name{revdbayes-internal}
\alias{revdbayes-internal}
\alias{process_data}
\alias{create_ru_list}
\alias{set_which_lam}
\alias{set_range_phi}
\alias{box_cox}
\alias{box_cox_vec}
\alias{box_cox_deriv}
\alias{check_sample_size}
\alias{check_sample_size_message}
\alias{logNegNA}
\title{Internal revdbayes functions}
\usage{
process_data(model, data, thresh, noy, use_noy, ros, weights = NULL)

create_ru_list(model, trans, rotate, min_xi, max_xi)

set_which_lam(model)

set_range_phi(model, phi_mid, se_phi, mult)

box_cox(x, lambda = 1, gm = 1, lambda_tol = 1e-06, poly_order = 3)

box_cox_vec(x, lambda = 1, lambda_tol = 1e-06)

box_cox_deriv(x, lambda = 1, lambda_tol = 1e-06, poly_order = 3)

check_sample_size(prior_name, n_check)

check_sample_size_message(prior_name, n_check)

logNegNA(x)
}
\description{
Internal revdbayes functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
