% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprior.R
\name{rprior_quant}
\alias{rprior_quant}
\title{Prior simulation of GEV parameters - prior on quantile scale}
\usage{
rprior_quant(n, shape, scale, prob)
}
\arguments{
\item{n}{A numeric scalar. The size of posterior sample required.}

\item{shape}{A numeric vector of length 3. Respective shape parameters of
the gamma priors for the quantile differences.}

\item{scale}{A numeric vector of length 3. Respective scale parameters of
the gamma priors for the quantile differences.}

\item{prob}{A numeric vector of length 3. Exceedance probabilities
corresponding to the quantiles used to specify the prior distribution.}
}
\value{
An \code{n} by 3 numeric matrix.
}
\description{
Simulates from the prior distribution for GEV parameters proposed in
Coles and Tawn (1996), based on independent gamma priors for differences
between quantiles.
}
\details{
The simulation is based on the way that the prior is constructed.
  [See Coles and Tawn (1996), Stephenson (1996) or the evdbayes user guide
  for details of the construction of the prior.] First, the quantile
  differences are simulated from the specified gamma distributions.
  Then the simulated quantiles are calculated. Then the GEV location,
  scale and shape parameters that give these quantile values are found,
  by solving numerically a set of three non-linear equations in which the
  GEV quantile function is equated to the simulated quantiles.  This is
  reduced to a one-dimensional optimisation over the GEV shape parameter.
}
\examples{
my_sh <- c(38.9,7.1,47)
my_sc <- c(1.5,6.3,2.6)
my_pr <- 10^-(1:3)
x1 <- rprior_quant(n = 1000, shape = my_sh, scale = my_sc, prob = my_pr)
}
\references{
Coles, S. G. and Tawn, J. A. (1996) A Bayesian analysis of
  extreme rainfall data. \emph{Appl. Statist.}, \strong{45}, 463-478.
  \url{http://dx.doi.org/10.2307/2986068}.

Stephenson, A. 2016. Bayesian Inference for Extreme Value
  Modelling. In \emph{Extreme Value Modeling and Risk Analysis: Methods and
  Applications}, edited by D. K. Dey and J. Yan, 257-80. London:
  Chapman and Hall. \url{http://dx.doi.org/10.1201/b19721-14}
}
\seealso{
\code{\link[evdbayes]{prior.quant}} to set this prior using the
  evdbayes package.

\code{\link[evdbayes]{posterior}}: evdbayes function that can sample from the
  prior distribution (using MCMC) if the argument \code{lh = "none"} is
  given.

\code{\link{rpost}} for sampling from an extreme value posterior
  distribution.
}

