% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_bin_prior}
\alias{set_bin_prior}
\title{Construction of a prior distribution for a binomial probability \eqn{p}}
\usage{
set_bin_prior(prior = c("jeffreys", "laplace", "haldane", "beta", "mdi"), ...)
}
\arguments{
\item{prior}{Either
\itemize{
  \item {A function that returns the value of the prior for \eqn{p}, or}
  \item {A character string giving the name of the prior for \eqn{p}.
    See \strong{Details} for a list of priors available for \eqn{p}.}
}}

\item{...}{Further arguments to be passed to the user-supplied or
in-built prior function.  For details of the latter see \strong{Details}.}
}
\value{
A list of class \code{"binprior"}.  The first component is the
  input prior, i.e. either the name of the prior or a user-supplied
  function.  The remaining components contain the numeric values of any
  hyperparameters in the prior.
}
\description{
Constructs a prior distribution for use as the argument \code{bin_prior} in
\code{\link{rpost}} or in \code{\link{binpost}}.  The user can either
specify their own prior function and arguments for hyperparameters or choose
from a list of in-built priors.
}
\details{
\strong{Binomial priors.} The names of the binomial priors set using
  \code{bin_prior} and details of hyperparameters are:
\itemize{
  \item {\code{"jeffreys"}: the \emph{Jeffreys} beta(1/2, 1/2) prior.}
  \item {\code{"laplace"}: the \emph{Bayes-Laplace} beta(1, 1) prior.}
  \item{\code{"haldane"}: the \emph{Haldane} beta(0, 0) prior.}
  \item{\code{"beta"}: a beta(\eqn{\alpha, \beta}) prior.  The argument
    \code{ab} is a vector containing \code{c}(\eqn{\alpha, \beta}).}
    The default is \code{ab = c(1, 1)}.
  \item{\code{"mdi"}: the prior
    \eqn{\pi(p) = 1.6186 p^p (1-p)^(1-p),     0 < p < 1.}
  }
}
}
\examples{
bp <- set_bin_prior(prior = "jeffreys")
}
\seealso{
\code{\link{binpost}} for sampling from a binomial posterior
  distribution.
}

