% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdbayes.R
\docType{data}
\name{venice}
\alias{venice}
\title{Largest Sea Levels in Venice}
\format{
A data frame with 51 rows and 10 columns.
}
\source{
Smith, R. L. (1986) Extreme value theory based on the \emph{r}
  largest annual events. \emph{Journal of Hydrology}, \strong{86}, 27-43.
  \url{https://doi.org/10.1016/0022-1694(86)90004-1}
}
\usage{
venice
}
\description{
The \code{venice} data frame has 51 rows and 10 columns. The jth column
contains the jth largest sea levels in Venice, for the years 1931-1981. Only
the largest six measurements are available for the year 1935; the
corresponding row contains four missing values. The years for each set of
measurements are given as row names.
}
\references{
Coles, S. G. (2001) \emph{An Introduction to Statistical
  Modelling of Extreme Values}. London: Springer.
  \url{https://doi.org/10.1007/978-1-4471-3675-0}
}
\keyword{datasets}
